Split Calibration module into per-type
